moved command to az partnercenter marketplace bundle
