+++ partnercenter/azext_partnercenter/operations/marketplace_bundle/commands.py
@@ -9,9 +9,9 @@ from azext_partnercenter._client_factory import cf_partnercenter
 
 
 def load_command_table(commands_loader, _):
-    custom_command_type = CliCommandType(operations_tmpl='azext_partnercenter.operations.marketplace_offer_bundle.custom#{}', client_factory=cf_partnercenter)
+    custom_command_type = CliCommandType(operations_tmpl='azext_partnercenter.operations.marketplace_bundle.custom#{}', client_factory=cf_partnercenter)
 
-    with commands_loader.command_group('partnercenter marketplace offer bundle', custom_command_type=custom_command_type, is_preview=True) as g:
+    with commands_loader.command_group('partnercenter marketplace bundle', custom_command_type=custom_command_type, is_preview=True) as g:
         g.custom_command('verify', 'verify_bundle', supports_no_wait=True, table_transformer=None)
         g.custom_command('build', 'build_bundle', supports_no_wait=True, table_transformer=None)
         g.custom_command('delete', 'delete_bundle', confirmation=True, supports_no_wait=True)
